Time: 14 ms (97.77%), Space: 11.7 MB (40.29%) - LeetHub

+++ 252-meeting-rooms/252-meeting-rooms.cpp
@@ -8,7 +8,7 @@ public:
             return true;
         }
         sort(intervals.begin(),intervals.end());
-        for(size_t i=0;i<intervals.size()-1;i++)
+        for(int i=0;i<intervals.size()-1;i++)
         {
             if(intervals[i][1] > intervals[i+1][0])
             {
